Add script to link images in sequence
ffmpeg requires that our files are named with in an unbroken numeric
sequence in the names in order to pass in a printf style pattern as the
input. This script takes a source and destination directory and creates
links in the destination for each image in the source. Destination files
are named "00001.jpg, 00002.jpg, ..." etc. so that the files have the
same sort order as the source directory.

Script is an adaptation of code found here:
http://programmer-art.org/articles/tutorials/ffmpeg-time-lapse