Modify signal handler to deal with SIGTERM as well
The System V init script we just added uses `kill $pid` to
terminate the take-photos.py script, which sends it a SIGTERM. As
a result, we now handle SIGTERM the same way as SIGINT so that we
shut down cleanly in both cases.

NOTE: there is a delay before the script exits because it has to
come out of its most recent sleep and hit the loop guard again
before it exits. This is reported by the system, however logs show
that we still get a clean exit. Would be nice to deal with this
and get out of the sleep early if possible.